+++ src/utils/index.ts
@@ -10,3 +10,4 @@ export * from './jobLogs';
 export * from './numbers';
 export * from './sorting';
 export * from './nodes';
+export * from './filtering';